#1 grant flows refactoring

+++ iam-web/src/main/java/org/uengine/iam/oauthtoken/OauthTokenService.java
@@ -10,6 +10,6 @@ public interface OauthTokenService {
             OauthClient oauthClient,
             OauthAccessToken accessToken,
             String claimJson,
-            Integer lifetime,
+            long lifetime,
             String type) throws Exception;
 }
