+++ iam-web/src/main/java/org/uengine/iam/oauthtoken/OauthTokenServiceImpl.java
@@ -60,7 +60,7 @@ public class OauthTokenServiceImpl implements OauthTokenService {
             OauthClient oauthClient,
             OauthAccessToken accessToken,
             String claimJson,
-            Integer lifetime,
+            long lifetime,
             String type) throws Exception {
 
         //발급 시간
